[Added] tests for decext
