--- internal/decext/copy.go
@@ -1,7 +1,0 @@
-package decext
-
-import "github.com/ericlagergren/decimal"
-
-func Copy(d *decimal.Big) *decimal.Big {
-	return new(decimal.Big).Set(d)
-}
